requirements.txt
paddlepaddle==1.7.2
# paddle-fl== 0.2.0
six>= 1.10.0
protobuf>= 3.1.0
zmq
tqdm
docstring_parser@ http://github.com/willthefrog/docstring_parser/tarball/master
typeguard; python_version >= '3.4'
tb-paddle
tensorboard>= 1.15
cython
pycocotools
paramiko~=2.7.2
gevent
loguru~=0.5.3
grpcio
aiohttp~=3.7.2

PyYAML~=5.3.1
numpy~=1.19.4
setuptools~=41.2.0
click~=7.1.2
attrs
